+++ lib/nsdata.py
