+++ lib/phot.py
