+++ lib/spec.py
